flex and rename
