+++ src/App.module.scss
@@ -1,3 +1,5 @@
 .App {
+  display: flex;
+  flex-direction: column;
   text-align: center;
 }
