--- src/Site/Contact.module.scss
@@ -1,2 +1,0 @@
-.contact {
-}
