--- src/Site/Header.module.scss
@@ -1,2 +1,0 @@
-.header {
-}
