--- src/Site/Projects.module.scss
@@ -1,2 +1,0 @@
-.projects {
-}
