--- src/Site/Skills.module.scss
@@ -1,2 +1,0 @@
-.skills {
-}
