+++ src/Website/Contact.module.scss
@@ -1,0 +1,4 @@
+.contact {
+  flex: 1;
+  border: 2px solid black;
+}
