+++ src/Website/Header.module.scss
@@ -1,0 +1,4 @@
+.header {
+  flex: 1;
+  border: 2px solid black;
+}
