+++ src/Website/Header.tsx
@@ -1,7 +1,11 @@
 import css from './Header.module.scss'
 
 const Header = (): JSX.Element => {
-  return <div className={css.header}></div>
+  return (
+    <div className={css.header}>
+      <h1>Header</h1>
+    </div>
+  )
 }
 
 export default Header
