+++ src/Website/Projects.module.scss
@@ -1,0 +1,4 @@
+.projects {
+  flex: 1;
+  border: 2px solid black;
+}
