+++ src/Website/Skills.module.scss
@@ -1,0 +1,4 @@
+.skills {
+  flex: 1;
+  border: 2px solid black;
+}
